improve events example
