+++ examples/snippets/package.json
@@ -9,5 +9,8 @@
     "stripe": "file:../..",
     "ts-node": "^10.9.2",
     "typescript": "^5.6.2"
+  },
+  "devDependencies": {
+    "@types/express": "4"
   }
 }
